refactor: simplify bounding box label drawing by removing font parameter and unused functions
